% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nn_viz.R
\name{nn_viz}
\alias{nn_viz}
\title{Neural Network Dimensionality Reduction Visualization}
\usage{
nn_viz(X, y, dim = 2, center.scale = TRUE, asp.equal = TRUE, views = 1)
}
\arguments{
\item{X}{A numeric matrix (n x m) representing the input features, where n is the number of samples and m is the number of features.}

\item{y}{A vector of length n representing the categories or labels corresponding to the input data.}

\item{dim}{Integer indicating the desired dimensionality of the visualization: 2 for 2D, 3 for 3D. Default is 2.}

\item{center.scale}{A logical (boolean) value indicating whether the data should be centered and scaled before processing.}

\item{asp.equal}{A logical (boolean) value, relevant only for 2D visualization, indicating whether the aspect ratio on both axes should be the same.}

\item{views}{An integer specifying the number of complementary projections. For 3D visualization, the maximum is 4. Subsequent views are based on orthogonal projections to capture different perspectives of the data.}
}
\value{
A list containing:
\describe{
  \item{projected_data}{The data projected onto the lower-dimensional space using the neural network.}
  \item{transformation_matrix}{The weight matrix of the neural network's first layer used for dimensionality reduction.}
}
Additionally, a Plotly plot representing the data points in the reduced feature space,
with points colored by their categories is displayed.
}
\description{
Creates a 2D (or 3D) visualization of multidimensional labeled data using a simple
neural network.
}
\examples{
data(iris)
X <- iris[,-5]
y <- iris[,5]
nn_viz(X, y)          # Default 2D visualization
nn_viz(X, y, dim = 3) # 3D visualization

}
