% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_viz.R
\name{pca_viz}
\alias{pca_viz}
\title{PCA Visualization}
\usage{
pca_viz(X, y, dim = 2, center.scale = TRUE)
}
\arguments{
\item{X}{A matrix representing the input features (quantitative variables).}

\item{y}{A vector representing the categories corresponding to the input data.}

\item{dim}{Integer indicating the desired dimensionality of the visualization: 2 for 2D, 3 for 3D. Default is 2.}

\item{center.scale}{A logical (boolean) value indicating whether the data should be centered and scaled before processing.}
}
\value{
\describe{
  \item{projected_data}{The data projected onto the principal components.}
  \item{transformation_matrix}{The eigenvectors corresponding to the selected principal components.}
}
Additionally, a plotly object representing the PCA visualization is displayed.
}
\description{
Creates a 2D or 3D visualization of multidimensional data using Principal Component Analysis (PCA).
}
\examples{
data(iris)
X <- iris[,-5]
y <- iris[,5]
pca_viz(X, y)          # Default 2D visualization
pca_viz(X, y, dim = 3) # 3D visualization
}
