\name{vizplore}
\alias{vizplore}
\docType{package}
\title{Visualization Package for Dimensionality Reduction}
\description{
  The `vizplore` package provides functions for dimensionality reduction and
  visualization of high-dimensional data using techniques like PCA, t-SNE, Canonical Correlation Analysis for categorical data, and neural networks.
  It supports 2D and 3D visualizations and is designed to help explore data patterns visually.
}
\details{
  This package includes the following key functions:
  \itemize{
    \item \code{\link{pca_viz}}: PCA-based dimensionality reduction and visualization.
    \item \code{\link{tsne_viz}}: t-SNE-based dimensionality reduction and visualization.
    \item \code{\link{catcca_viz}}: Dimensionality reduction and visualization based on Canonical Correlation Analysis for categorical data.
    \item \code{\link{nn_viz}}: Neural network-based dimensionality reduction and visualization.
  }
}
\author{Your Name}
\maintainer{Your Name <your.email@example.com>}
\license{GPL-3}
