% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cca_viz.R
\name{cca_viz}
\alias{cca_viz}
\title{CCA Visualization}
\usage{
cca_viz(X, y, dim = 2, center.scale = TRUE, asp.equal = TRUE, views = 1)
}
\arguments{
\item{X}{A matrix representing the input features (quantitative variables).}

\item{y}{A vector representing the categories corresponding to the input data.}

\item{dim}{Integer indicating the desired dimensionality of the visualization: 2 for 2D, 3 for 3D. Default is 2.}

\item{center.scale}{A logical (boolean) value indicating whether the data should be centered and scaled before processing.}

\item{asp.equal}{A logical (boolean) value, relevant only for 2D visualization, indicating whether the aspect ratio on both axes should be the same.}

\item{views}{An integer specifying the number of complementary projections. For 3D visualization, the maximum is 4. Subsequent views are based on orthogonal projections to capture different perspectives of the data.}

\item{center}{A logical value indicating whether to center the quantitative data.}
}
\value{
A list containing:
\describe{
  \item{projected_data}{The projected data onto the canonical components.}
  \item{transformation_matrix}{The eigenvectors corresponding to the canonical components.}
}
Additionally, a Plotly plot representing the data points in the reduced feature space,
with points colored by their categories is displayed.
}
\description{
Creates a 2D (or 3D) visualization of multidimensional labeled data using
categorical canonical correlation analysis.
}
\examples{
data(iris)
X <- iris[,-5]
y <- iris[,5]
cca_viz(X, y)          # Default 2D visualization
cca_viz(X, y, dim = 3) # 3D visualization

}
